fix bar computation
Compute bars corrrectly, instead of skipping the computation when the
cell above or to the left is zero laps down. Further, draw a bar only
when the other cell _is_ zero laps down; drawing bars between cars that
are down different numbers of laps _sounds_ nice but actually looks
terrible.