rework for multiple windows
Add support for having multiple windows at once, including "New Window"
and "Close Window" commands. The default sample data is displayed only
in the initial window created when the program is started.

This temporarily disables the ability to read data from a file and
display it in a window, pending future rework to handle this case.