Improve file-loading abstraction
Move the file-loading code into a separate module, add support for
reading Orbits passings in CSV format, and add a wrapper that detects
the file type and calls the appropriate loader.

Additionally, add an optional 'files' argument to the top-level GUI
class. If a list of filenames is given, windows are created for each
file listed. The main program now uses this to allow opening multiple
files specified on the command line.