Complete zoom support
When zooming, actually change the font sizes. Also adjust the scroll
positions so that zooming doesn't change what part of the chart you're
looking at (specifically, the upper-left corner is fixed).

As an extra bonus, lap and column headers are now shown in a bold font,
and the bars get slightly thicker as the zoom level increases.