ColorListWidget; avoid modifying the color list
In 'single' mode, ColorListWidget displays only one color swatch and
only permits editing of that single color, even if the passed-in color
list contained more than one color. However, that doesn't really give it
license to _modify_ the passed-in list, and it can't maintain a private
copy, since changing that color need to take effect outside the widget.
So, stop truncating the passed in list, but if the single flag was set,
behave as if there were only one color in the list.